[REEF-143] Refactor REEF.Examples into one project
This is to consolidate all reef examples into one project
Updated folder structures and name spaces
Updated references
Updated pom file for building Org.Apache.REEF

JIRA: Reef-143. (https://issues.apache.org/jira/browse/REEF-143)
Author: Julia Wang  Email: jwang98052@yahoo.com
