+++ lang/cs/Org.Apache.REEF.Examples/Org.Apache.REEF.Examples.csproj
@@ -23,8 +23,8 @@ under the License.
     <ProjectGuid>{75503F90-7B82-4762-9997-94B5C68F15DB}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>Org.Apache.REEF.Examples.Tasks</RootNamespace>
-    <AssemblyName>Org.Apache.REEF.Examples.Tasks</AssemblyName>
+    <RootNamespace>Org.Apache.REEF.Examples</RootNamespace>
+    <AssemblyName>Org.Apache.REEF.Examples</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <RestorePackages>true</RestorePackages>
@@ -79,14 +79,34 @@ under the License.
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="FailedTask\FailedTask.cs" />
-    <Compile Include="HelloTask\HelloService.cs" />
-    <Compile Include="HelloTask\HelloTask.cs" />
-    <Compile Include="HelloTask\HelloTaskMessage.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\AnotherHelloAllocatedEvaluatorHandler.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloActiveContextHandler.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloAllocatedEvaluatorHandler.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloCompletedEvaluatorHandler.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloDriverRestartActiveContextHandler.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloDriverRestartRunningTaskHandler.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloEvaluatorRequestorHandler.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloFailedEvaluatorHandler.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloFailedTaskHandler.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloHttpHandler.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloRestartHandler.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloRunningTaskHandler.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloSimpleEventHandlers.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloStartHandler.cs" />
+    <Compile Include="HelloCLRBridge\Handlers\HelloTaskMessageHandler.cs" />
+    <Compile Include="HelloCLRBridge\HelloTraceListener.cs" />
+    <Compile Include="RetainedEvalCLRBridge\Handlers\RetainedEvalActiveContextHandler.cs" />
+    <Compile Include="RetainedEvalCLRBridge\Handlers\RetainedEvalAllocatedEvaluatorHandler.cs" />
+    <Compile Include="RetainedEvalCLRBridge\Handlers\RetainedEvalEvaluatorRequestorHandler.cs" />
+    <Compile Include="RetainedEvalCLRBridge\Handlers\RetainedEvalStartHandler.cs" />
+    <Compile Include="Tasks\FailedTask\FailedTask.cs" />
+    <Compile Include="Tasks\HelloTask\HelloService.cs" />
+    <Compile Include="Tasks\HelloTask\HelloTask.cs" />
+    <Compile Include="Tasks\HelloTask\HelloTaskMessage.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ShellTask\ShellTask.cs" />
-    <Compile Include="StreamingTasks\StreamTask1.cs" />
-    <Compile Include="StreamingTasks\StreamTask2.cs" />
+    <Compile Include="Tasks\ShellTask\ShellTask.cs" />
+    <Compile Include="Tasks\StreamingTasks\StreamTask1.cs" />
+    <Compile Include="Tasks\StreamingTasks\StreamTask2.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="$(SolutionDir)\Org.Apache.REEF.Tang\Org.Apache.REEF.Tang.csproj">
@@ -104,6 +124,14 @@ under the License.
     <ProjectReference Include="$(SolutionDir)\Org.Apache.REEF.Driver\Org.Apache.REEF.Driver.csproj">
       <Project>{a6baa2a7-f52f-4329-884e-1bcf711d6805}</Project>
       <Name>Org.Apache.REEF.Driver</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Org.Apache.REEF.Network\Org.Apache.REEF.Network.csproj">
+      <Project>{883ce800-6a6a-4e0a-b7fe-c054f4f2c1dc}</Project>
+      <Name>Org.Apache.REEF.Network</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\Org.Apache.REEF.Wake\Org.Apache.REEF.Wake.csproj">
+      <Project>{cdfb3464-4041-42b1-9271-83af24cd5008}</Project>
+      <Name>Org.Apache.REEF.Wake</Name>
     </ProjectReference>
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
